Rename HeaderImage to Header, and Image to HeaderImage, update imports and calls
