+++ src/components/App/App.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import Card from '@material-ui/core/Card';
-import HeaderImage from '../HeaderImage/HeaderImage';
+import Header from '../Header/Header';
 import TodoList from '../TodoList/TodoList';
 import rainbow from '../../images/rainbow.png';
 import tree from '../../images/pine.png';
@@ -87,7 +87,7 @@ const App = ({ todos }) => {
     <div className={backgroundChoice}>
       <div className={styles.app}>
         <Card className={styles.root}>
-          <HeaderImage
+          <Header
             items={items}
             isFormDisplay={isFormDisplay}
             toggleForm={toggleForm}
