+++ src/components/Header/Header.module.css
@@ -1,0 +1,12 @@
+.darkShade {
+  position: absolute;
+  background-color: rgba(158, 99, 217, 0.4);
+  width: 30%;
+  height: 273px;
+  border-radius: 5px;
+  left: 70%;
+}
+.headerImage {
+  width: 480px;
+  height: 270px;
+}
