--- src/components/Image/Image.module.css
@@ -1,8 +1,0 @@
-.image {
-  border-radius: 4px;
-  width: 100%;
-  position: absolute;
-  left: 0%;
-  height: 273px;
-  object-fit: cover;
-}
